add test scenario to check unsubscribing from playback info events
