+++ tests/bdd/test_media_renderer_api.py
@@ -1,8 +1,6 @@
 from pytest_bdd import scenarios, given, when, then, parsers, scenario
-import pytest
 from .async_utils import sync
 from functools import wraps
-import asyncio
 import logging
 from .api_utils import get_playback_info_path
